indices from u32 to u16

+++ src/render/model.rs
@@ -4,8 +4,8 @@ use super::{buffer::DynamicBuffer, Vertex};
 /// Represents a mesh that has been sent to the GPU.
 pub struct Model<V: Vertex>{
     vbuf: Buffer<V>,
-    ibuf: Buffer<u32>,
-    pub num_indices: u32,
+    ibuf: Buffer<u16>,
+    pub num_indices: u16,
 }
 
 impl<V: Vertex> Model<V>{
@@ -20,22 +20,22 @@ impl<V: Vertex> Model<V>{
         Some(Self {
             vbuf,
             ibuf,
-            num_indices: mesh.indices().len() as u32,
+            num_indices: mesh.indices().len() as u16,
         })
     }
 
     
     pub fn vbuf(&self) -> &wgpu::Buffer { &self.vbuf.buff }
     pub fn ibuf(&self) -> &wgpu::Buffer { &self.ibuf.buff }
-    pub fn len(&self) -> u32 { self.vbuf.len() as u32}
+    pub fn len(&self) -> u16 { self.vbuf.len() as u16}
 }
 
 
 /// Represents a mesh that has been sent to the GPU.
 pub struct DynamicModel<V: Vertex> {
     vbuf: DynamicBuffer<V>,
-    ibuf: DynamicBuffer<u32>,
-    pub num_indices: u32,
+    ibuf: DynamicBuffer<u16>,
+    pub num_indices: u16,
 }
 
 impl<V: Vertex> DynamicModel<V> {
@@ -50,7 +50,7 @@ impl<V: Vertex> DynamicModel<V> {
     pub fn update(&mut self, queue: &wgpu::Queue, mesh: &Mesh<V>, offset: usize) {
         self.vbuf.update(queue, mesh.vertices(), offset);
         self.ibuf.update(queue, mesh.indices(), offset);
-        self.num_indices = mesh.indices().len() as u32;
+        self.num_indices = mesh.indices().len() as u16;
     }
 
     pub fn vbuf(&self) -> &wgpu::Buffer { &self.vbuf.buff }
